Fixes to separate region instance group specific attributes.
